--- app/src/main/java/com/github/elianaferreira/movieslist/models/ApiKey.kt
@@ -1,10 +1,0 @@
-package com.github.elianaferreira.movieslist.models
-import androidx.annotation.Keep
-
-import com.google.gson.annotations.SerializedName
-
-@Keep
-data class ApiKey(
-    @SerializedName("apiKey")
-    val apiKey: String
-)
