+++ app/src/main/java/com/github/elianaferreira/movieslist/stories/list/ListActivity.kt
@@ -1,4 +1,4 @@
-package com.github.elianaferreira.movieslist.list
+package com.github.elianaferreira.movieslist.stories.list
 
 import android.app.SearchManager
 import android.content.Context
@@ -14,10 +14,9 @@ import androidx.appcompat.widget.Toolbar
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.github.elianaferreira.movieslist.R
-import com.github.elianaferreira.movieslist.activities.MovieDetailActivity
-import com.github.elianaferreira.movieslist.activities.TVShowDetailActivity
-import com.github.elianaferreira.movieslist.adapters.MoviesAdapter
-import com.github.elianaferreira.movieslist.home.Category
+import com.github.elianaferreira.movieslist.stories.detail.MovieDetailActivity
+import com.github.elianaferreira.movieslist.stories.detail.TVShowDetailActivity
+import com.github.elianaferreira.movieslist.stories.home.Category
 import com.github.elianaferreira.movieslist.utils.RequestManager
 import com.github.elianaferreira.movieslist.utils.Utils
