+++ app/src/main/java/com/github/elianaferreira/movieslist/utils/RequestManager.kt
@@ -11,7 +11,7 @@ import com.android.volley.toolbox.JsonObjectRequest
 import com.android.volley.toolbox.Volley
 import com.github.elianaferreira.movieslist.BuildConfig
 import com.github.elianaferreira.movieslist.R
-import com.github.elianaferreira.movieslist.list.MoviesList
+import com.github.elianaferreira.movieslist.stories.list.MoviesList
 import com.github.elianaferreira.movieslist.models.*
 import com.google.gson.Gson
 import java.lang.StringBuilder
