+++ app/src/main/res/layout/activity_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".list.ListActivity"
+    tools:context=".stories.list.ListActivity"
     android:background="@color/white">
 
     <include
