+++ app/src/main/res/layout/activity_movie_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".activities.MovieDetailActivity">
+    tools:context=".stories.detail.MovieDetailActivity">
 
     <ScrollView
         android:layout_width="match_parent"
