+++ app/src/main/res/layout/activity_tvshow_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".activities.TVShowDetailActivity">
+    tools:context=".stories.detail.TVShowDetailActivity">
 
     <ScrollView
         android:layout_width="match_parent"
